chore: test and improvements
