+++ python/letsql/examples/tests/test_examples.py
@@ -1,0 +1,17 @@
+import pytest
+
+import letsql as ls
+from letsql.examples import Example
+from letsql.examples.core import whitelist
+
+
+def test_whitelist():
+    assert dir(ls.examples) == whitelist
+
+
+def test_attributes():
+    example = ls.examples.astronauts
+    assert isinstance(example, Example)
+
+    with pytest.raises(AttributeError):
+        ls.examples.missing
